Started on the final project code
I started the base of the final project code. Nothing works together yet. The right button capturing should be finished.

My plan for the solution is to run a couple things at the same time, first is finding the target pattern, and the other is determining which of the 8 sectors the robot is located in. 

To find the target pattern the robot will use the black lines on the ground to turn and approach the pattern, then scan it and compare it to the target pattern, it will then backup, turn and move clockwise to the next pattern.

To determine the sector the robot will need to keep track of the color of the ground below it and use that to determine which sector it is in as it travels. The lab code can be very useful for this because it does something very similar.