Begining work on localization
In the middle of implementing the localization function. Currently set up to just detect the colour underneath it then the robot needs to be manually moved to the next section. Localization does not function properly. Moving clockwise around the arena in implemented but currently commented out.

Many of the simple requirements have also been implemented such as the 4 minute timer and the initial LED settings.