Added localization and target recognition
The localization works quite well. I made sure to not use 1 and 0 in order to provide a level of uncertainty to it. Target recognition means that the robot knows to stop once it reaches the pattern it is searching for. The robot still needs to know how to return to the center of the course. Also, the robot consistently skips the black bar on tile 1 (the top middle tile). I'm not sure why.