It's done
I don't think there is anything else that needs to be added. The robot is now fully capable of selecting a pattern, searching around the arena for the pattern, localizing itself, and returning to the center square with more than 3/4 of its body being inside of it. If something is missing, feel free to adjust the code.